move `isProposedApiEnabled` and `checkProposedApiEnabled`-util back to workbench-layer
